requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-extensions==2.2.5
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
